+++ R/build_sql.r
@@ -1,0 +1,7 @@
+
+topolclause <- function(x) {
+  switch(x, 
+         area = "IsArea([ID])", 
+         line = "IsLine([ID])", 
+         point = "IsPoint([ID])")
+}