+++ NEWS.md
@@ -1,8 +1,26 @@
-# manifoldr 0.0.3.9001
 
-* preparations for CRAN
+# manifoldr dev
+
+* preparation for CRAN
+
+# manifoldr 0.0.5
+
+* added logic for WHERE argument to add clauses with AND to internal readmfd
+
+
+# manifoldr 0.0.4
+
+* Added `Surface` and `Image` function to read surfaces as `Rasterlayer` and images as RGB `RasterBrick`. 
 
 * Added a `NEWS.md` file to track changes to the package.
 
+* Datasets addded to extdata `MONTARA_20m.map`, `V20160202016022.L3m_R3QL_NPP_CHL_chlor_a_9km.map`
 
+* Added ability to read CRS information converted to PROJ.4
+
+* Added ability to read from drawings as 'Spatial' classes
+
+# manifoldr 0.0.3 and earlier
+
+* initial builds